requirements.txt
fastapi
uvicorn
opencv-python
google-genai
python-dotenv
pydantic
